requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
openai>=1.40.0
python-dotenv>=1.0.0
gradio>=4.44.0
jinja2>=3.1.0
matplotlib>=3.8.0
